decoder , ipports  int-> int8_t

+++ spi/spi_master_slave/spi_master/utilities/base_types/Decoder.hpp
@@ -7,9 +7,9 @@
 class Decoder
 {
 public:
-  Decoder(int a, int b, int c);
+  Decoder(uint8_t a, uint8_t b, uint8_t c);
 
-  void activePort(int port = 0);
+  void activePort(uint8_t port = 0);
 
 private:
   OutputPort portA;
